--- src/main/java/com/jfrog/ide/common/components/ComparableSeverityTreeNode.java
@@ -1,13 +1,0 @@
-package com.jfrog.ide.common.components;
-
-import com.jfrog.ide.common.components.subentities.Severity;
-
-import javax.swing.tree.DefaultMutableTreeNode;
-
-public abstract class ComparableSeverityTreeNode extends DefaultMutableTreeNode implements Comparable<ComparableSeverityTreeNode> {
-    public abstract Severity getSeverity();
-
-    public int compareTo(ComparableSeverityTreeNode other) {
-        return other.getSeverity().ordinal() - this.getSeverity().ordinal();
-    }
-}
