+++ src/main/java/com/jfrog/ide/common/nodes/FileTreeNode.java
@@ -1,12 +1,10 @@
-package com.jfrog.ide.common.components;
+package com.jfrog.ide.common.nodes;
 
-import com.jfrog.ide.common.components.subentities.Severity;
-import org.apache.commons.collections.CollectionUtils;
+import com.jfrog.ide.common.nodes.subentities.Severity;
 
 import java.io.File;
-import java.util.Comparator;
 
-public class FileTreeNode extends ComparableSeverityTreeNode implements SubtitledTreeNode {
+public class FileTreeNode extends SortableChildrenTreeNode implements SubtitledTreeNode, Comparable<FileTreeNode> {
     protected String fileName;
     protected String filePath;
     protected Severity topSeverity = Severity.Normal;
@@ -52,9 +50,8 @@ public class FileTreeNode extends ComparableSeverityTreeNode implements Subtitle
         }
     }
 
-    public void sortChildren() {
-        if (CollectionUtils.isNotEmpty(children)) {
-            children.sort(Comparator.comparing(treeNode -> ((ComparableSeverityTreeNode) treeNode)));
-        }
+    @Override
+    public int compareTo(FileTreeNode other) {
+        return other.getSeverity().ordinal() - this.getSeverity().ordinal();
     }
 }
