+++ src/main/java/com/jfrog/ide/common/nodes/VulnerabilityNode.java
@@ -1,15 +1,15 @@
-package com.jfrog.ide.common.components;
+package com.jfrog.ide.common.nodes;
 
-import com.jfrog.ide.common.components.subentities.Cve;
-import com.jfrog.ide.common.components.subentities.ResearchInfo;
-import com.jfrog.ide.common.components.subentities.Severity;
+import com.jfrog.ide.common.nodes.subentities.Cve;
+import com.jfrog.ide.common.nodes.subentities.ResearchInfo;
+import com.jfrog.ide.common.nodes.subentities.Severity;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 
-import static com.jfrog.ide.common.components.subentities.Severity.getNotApplicableSeverity;
+import static com.jfrog.ide.common.nodes.subentities.Severity.getNotApplicableSeverity;
 
 /**
  * @author yahavi
