+++ src/main/java/com/jfrog/ide/common/nodes/subentities/License.java
@@ -1,4 +1,4 @@
-package com.jfrog.ide.common.components.subentities;
+package com.jfrog.ide.common.nodes.subentities;
 
 public class License {
     private final String name;
