+++ src/main/java/com/jfrog/ide/common/nodes/subentities/ResearchInfo.java
@@ -1,4 +1,4 @@
-package com.jfrog.ide.common.components.subentities;
+package com.jfrog.ide.common.nodes.subentities;
 
 public class ResearchInfo {
     private final Severity severity;
